+++ packages/core/transport/src/LoggerlessTransport.ts
@@ -8,6 +8,8 @@ import {
 
 /**
  * For implementing libraries that aren't logging libraries
+ *
+ * {@link https://loglayer.dev/transports/creating-transports.html | Creating Transports}
  */
 export abstract class LoggerlessTransport implements LogLayerTransport {
   /**